% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcgatools.r
\name{extract.cols.ix}
\alias{extract.cols.ix}
\title{Extract indexes of columns of interest}
\usage{
extract.cols.ix(tcga.data, prefix = ".+", suffix = ".+")
}
\arguments{
\item{tcga.data}{Dataframe of TCGA data}

\item{prefix}{Regular expression for a data type in the TCGA table (e.g. RNA, CNA)}

\item{suffix}{Regular expression for a gene defining a molecular class}
}
\value{
Returns vector of indexes of columns with prefix or suffix of interest
}
\description{
Extract the indexes of the columns with a given prefix (data type) and/or suffix (molecular class)
}
\examples{
data("tcgaData")
extract.cols.ix(example.tcga.table, prefix= 'RNA', suffix='PTEN')
## Use regular expressions to select multiple prefixes/suffixes
extract.cols.ix(example.tcga.table, prefix='RNA|CNA')
}
