% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcgatools.r
\name{extract.subtype}
\alias{extract.subtype}
\title{Extract columns of a subtype}
\usage{
extract.subtype(tcga.data, subtype, with_metadata = TRUE)
}
\arguments{
\item{tcga.data}{Dataframe of TCGA data}

\item{subtype}{Regular expression for data subtype (RNA, RNAZ, PPA)}

\item{with_metadata}{Boolean indicating wheather to include metadata columns in the output dataframe}
}
\value{
Returns TCGA dataframe with selected columns
}
\description{
Extracts columns of TCGA dataframe relative to a given data subtype (RNA, RNAZ, PPA)
}
\examples{
## Normalize columns containing RNA data
data('tcgaData')
extract.subtype(example.tcga.table, 'RNA')
## Use pattern matching to select multiple subtypes
extract.subtype(example.tcga.table, 'RNA|CNA')
}
