% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcgatools.r
\name{norm.columns}
\alias{norm.columns}
\title{Normalize selected columns}
\usage{
norm.columns(tcga.data, cols.to.normalize, show.plot = FALSE)
}
\arguments{
\item{tcga.data}{Dataframe of TCGA data}

\item{cols.to.normalize}{Index of columns to use for normalization}

\item{show.plot}{Boolean indicating wheather to show boxplots of distribution of in columns of interest before and after normalization}
}
\value{
Returns TCGA dataframe with selected columns normalized
}
\description{
Return a data frame with only selected columns normalized, based on median and mad
}
\examples{
## Normalize columns containing RNA data
data('tcgaData')
columns.of.interest <- extract.cols.ix(example.tcga.table, prefix='RNA')
norm.columns(example.tcga.table, cols.to.normalize = columns.of.interest)
}
