% Generated by roxygen2 (4.0.2): do not edit by hand
\name{parse_barcodes}
\alias{parse_barcodes}
\title{Parse TCGA barcodes}
\usage{
parse_barcodes(x, annotate = FALSE, verbose = FALSE)
}
\arguments{
\item{x}{character vector of barcodes}

\item{verbose}{report barcode type match similarities}
}
\description{
Extract and expand information embedded in TCGA sample barcodes. Details
about the barcode format is provided
\href{https://wiki.nci.nih.gov/display/TCGA/TCGA+Barcode}{here}.
}
\examples{
barcodes <- c('TCGA-EJ-7321-11A-01R-2263-07', 'TCGA-EJ-7321-11A-01R-2263-07')
parse_barcodes(barcodes, annotate = TRUE)
}

