% Generated by roxygen2 (4.0.2): do not edit by hand
\name{parse_barcodes}
\alias{parse_barcodes}
\title{Parse TCGA barcodes}
\usage{
parse_barcodes(x, annotations = "none")
}
\arguments{
\item{x}{character vector of barcodes}

\item{annotations}{character string indicating whether barcodes should be
  only be parsed (\code{annotations = "none"}), or annotated using either
  \code{"short"} or \code{"long"} values}
}
\description{
Extract and expand information embedded in TCGA sample barcodes. Details
about the barcode format is provided
\href{https://wiki.nci.nih.gov/display/TCGA/TCGA+Barcode}{here}.
}
\examples{
barcodes <- c('TCGA-10-7321-11A-01R-2263-07', 'TCGA-Q5-7321-01A-01D-2263-32')
parse_barcodes(barcodes, annotations = "long")
}

